Pull out more code from HomeScreenFragment
