+++ platform/android/app/src/main/AndroidManifest.xml
@@ -11,11 +11,6 @@
         android:networkSecurityConfig="@xml/network_security_config"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity
-            android:name=".HomeScreenFragment"
-            android:exported="false"
-            android:label="@string/title_activity_home_screen"
-            android:theme="@style/AppTheme" />
 
         <meta-data
             android:name="android.webkit.WebView.MetricsOptOut"
@@ -48,7 +43,7 @@
             android:theme="@style/Theme.AppCompat.Light.NoActionBar" />
 
         <activity
-            
+
             android:name=".settings.SettingsActivity"
             android:exported="false"
             android:label="@string/settings_title"
