+++ platform/android/app/src/main/kotlin/coop/polypoly/polypod/homescreen/HomeScreenFragment.kt
@@ -12,8 +12,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.asImageBitmap
-import androidx.compose.ui.graphics.toArgb
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.platform.LocalConfiguration
@@ -27,9 +25,7 @@ import androidx.fragment.app.Fragment
 import androidx.navigation.findNavController
 import androidx.navigation.fragment.findNavController
 import com.google.accompanist.flowlayout.FlowRow
-import coop.polypoly.polypod.HomeScreenFragmentDirections
 import coop.polypoly.polypod.R
-import coop.polypoly.polypod.luminance
 
 class HomeScreenFragment : Fragment() {
     private val viewModel = HomeScreenViewModel()
@@ -108,7 +104,7 @@ fun topBar(onInfoClick: () -> Unit, onSettingsClick: () -> Unit) {
 }
 
 @Composable
-fun screen(sectionModels: List<SectionModel>) {
+fun createScreen(sectionModels: List<SectionModel>) {
     val configuration = LocalConfiguration.current
     val tilesPerContainer = 3
 
@@ -164,10 +160,6 @@ fun screen(sectionModels: List<SectionModel>) {
         layout = screenLayout
     )
     return Screen(screen = screen)
-}
-
-fun isLight(color: Color): Boolean {
-    return luminance(color.toArgb()) > 100
 }
 
 @Composable
@@ -226,470 +218,6 @@ fun Section(section: Section) {
     }
 }
 
-@Composable
-fun LargeLeftContainerView(container: Container) {
-    Row(
-        horizontalArrangement = Arrangement.spacedBy(
-            container.layout.horizontalInterItemSpacing
-        )
-    ) {
-        BigTileView(
-            tile = container.tiles.first()
-        )
-        Column(
-            verticalArrangement = Arrangement.spacedBy(
-                container.layout.verticalInterItemSpacing
-            )
-        ) {
-            container.tiles.drop(1).forEach { tile ->
-                SmallTileView(
-                    tile = tile
-                )
-            }
-        }
-    }
-}
-
-@Composable
-fun LargeRightContainerView(container: Container) {
-    Row(
-        horizontalArrangement = Arrangement.spacedBy(
-            container.layout.horizontalInterItemSpacing
-        )
-    ) {
-        val layoutNumTilesThreshold = 3
-        if (container.tiles.count() < layoutNumTilesThreshold) {
-            container.tiles.forEach { tile ->
-                SmallTileView(tile = tile)
-            }
-        } else {
-            Column(
-                verticalArrangement = Arrangement.spacedBy(
-                    container.layout.verticalInterItemSpacing
-                )
-            ) {
-                container.tiles.dropLast(1).forEach { tile ->
-                    SmallTileView(
-                        tile = tile
-                    )
-                }
-            }
-            BigTileView(
-                tile = container.tiles.last()
-            )
-        }
-    }
-}
-
-@Composable
-fun RowContainerView(container: Container) {
-    FlowRow(
-        crossAxisSpacing = container.layout.horizontalInterItemSpacing,
-        mainAxisSpacing = container.layout.verticalInterItemSpacing
-    ) {
-        container.tiles.forEach { tile ->
-            when (tile.type) {
-                TileType.BIG -> BigTileView(
-                    tile = tile
-                )
-                TileType.MEDIUM -> MediumTileView(
-                    tile = tile
-                )
-                TileType.SMALL -> SmallTileView(
-                    tile = tile
-                )
-            }
-        }
-    }
-}
-
-@Composable
-fun BigTileView(tile: Tile) {
-    val foregroundColor = if (isLight(tile.model.backgroundColor)) Color.Black else Color.White // ktlint-disable max-line-length
-    Card(
-        modifier = Modifier
-            .width(tile.layout.width)
-            .height(tile.layout.height)
-            .clickable {
-                tile.model.onSelection()
-            },
-        shape = RoundedCornerShape(tile.layout.cornerRadius)
-    ) {
-        Column(
-            horizontalAlignment = Alignment.CenterHorizontally,
-            modifier = Modifier
-                .background(tile.model.backgroundColor)
-                .padding(
-                    top = tile.layout.topPadding,
-                    start = tile.layout.startPadding,
-                    end = tile.layout.endPadding,
-                    bottom = tile.layout.bottomPadding
-                )
-        ) {
-            tile.model.image?.also {
-                Image(
-                    bitmap = it.asImageBitmap(),
-                    contentDescription = null,
-                    // Takes all the height left after the text is placed
-                    modifier = Modifier.weight(1.0f),
-                    contentScale = ContentScale.Fit,
-                    alignment = Alignment.Center
-                )
-            }
-            Spacer(
-                modifier = Modifier.defaultMinSize(
-                    minHeight = tile.layout.verticalSpacing,
-                    minWidth = tile.layout.width
-                )
-            )
-            Column {
-                Text(
-                    text = tile.model.title,
-                    color = foregroundColor,
-                    overflow = TextOverflow.Ellipsis,
-                    fontWeight = tile.style.titleFont.weight,
-                    fontFamily = FontFamily(
-                        Font(tile.style.titleFont.family)
-                    ),
-                    fontSize = tile.style.titleFont.size,
-                    lineHeight = tile.style.titleFont.lineHeight,
-                    textAlign = tile.style.titleFont.alignment,
-                )
-                Spacer(
-                    modifier = Modifier.defaultMinSize(
-                        minHeight = tile.layout.textVerticalSpacing,
-                        minWidth = tile.layout.width
-                    )
-                )
-                Text(
-                    text = tile.model.description,
-                    color = foregroundColor,
-                    overflow = TextOverflow.Ellipsis,
-                    fontWeight = tile.style.descriptionFont!!.weight,
-                    fontFamily = FontFamily(
-                        Font(tile.style.descriptionFont.family)
-                    ),
-                    fontSize = tile.style.descriptionFont.size,
-                    lineHeight = tile.style.descriptionFont.lineHeight,
-                    textAlign = tile.style.descriptionFont.alignment,
-                )
-            }
-        }
-    }
-}
-
-@Composable
-fun MediumTileView(tile: Tile) {
-    val foregroundColor = if (isLight(tile.model.backgroundColor)) Color.Black else Color.White // ktlint-disable max-line-length
-    Card(
-        modifier = Modifier
-            .width(tile.layout.width)
-            .height(tile.layout.height)
-            .clickable {
-                tile.model.onSelection()
-            },
-        shape = RoundedCornerShape(tile.layout.cornerRadius)
-    ) {
-        Row(
-            modifier = Modifier
-                .background(tile.model.backgroundColor)
-        ) {
-            tile.model.image?.also {
-                Image(
-                    bitmap = it.asImageBitmap(),
-                    contentDescription = null,
-                    contentScale = ContentScale.Fit,
-                    alignment = Alignment.Center,
-                    modifier = Modifier.width(tile.layout.height)
-                )
-            }
-
-            Column(
-                modifier = Modifier.padding(
-                    top = tile.layout.textTopPadding,
-                    bottom = tile.layout.textBottomPadding,
-                    start = tile.layout.textStartPadding,
-                    end = tile.layout.textEndPadding
-                )
-            ) {
-                Text(
-                    text = tile.model.title,
-                    color = foregroundColor,
-                    overflow = TextOverflow.Ellipsis,
-                    fontWeight = tile.style.titleFont.weight,
-                    fontFamily = FontFamily(
-                        Font(tile.style.titleFont.family)
-                    ),
-                    fontSize = tile.style.titleFont.size,
-                    lineHeight = tile.style.titleFont.lineHeight,
-                    textAlign = tile.style.titleFont.alignment,
-                )
-                Spacer(
-                    modifier = Modifier.defaultMinSize(
-                        minHeight = tile.layout.textVerticalSpacing,
-                        minWidth = tile.layout.width
-                    )
-                )
-                Text(
-                    text = tile.model.description,
-                    color = foregroundColor,
-                    overflow = TextOverflow.Ellipsis,
-                    fontWeight = tile.style.descriptionFont!!.weight,
-                    fontFamily = FontFamily(
-                        Font(tile.style.descriptionFont.family)
-                    ),
-                    fontSize = tile.style.descriptionFont.size,
-                    lineHeight = tile.style.descriptionFont.lineHeight,
-                    textAlign = tile.style.descriptionFont.alignment,
-                )
-            }
-        }
-    }
-}
-
-@Composable
-fun SmallTileView(tile: Tile) {
-    val foregroundColor = if (isLight(tile.model.backgroundColor)) Color.Black else Color.White // ktlint-disable max-line-length
-    Card(
-        modifier = Modifier
-            .width(tile.layout.width)
-            .height(tile.layout.height)
-            .clickable {
-                tile.model.onSelection()
-            },
-        shape = RoundedCornerShape(tile.layout.cornerRadius)
-    ) {
-        Column(
-            modifier = Modifier
-                .background(tile.model.backgroundColor)
-                .padding(
-                    top = tile.layout.topPadding,
-                    start = tile.layout.startPadding,
-                    end = tile.layout.endPadding,
-                    bottom = tile.layout.bottomPadding
-                ),
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.SpaceBetween
-        ) {
-            tile.model.image?.also {
-                Image(
-                    bitmap = it.asImageBitmap(),
-                    contentDescription = null,
-                    // Takes all the height left after the text is placed
-                    modifier = Modifier.weight(1.0f),
-                    contentScale = ContentScale.Fit,
-                    alignment = Alignment.Center
-                )
-            }
-            Spacer(
-                modifier = Modifier.defaultMinSize(
-                    minWidth = tile.layout.width,
-                    minHeight = tile.layout.verticalSpacing
-                )
-            )
-            Text(
-                text = tile.model.title,
-                color = foregroundColor,
-                overflow = TextOverflow.Ellipsis,
-                fontWeight = tile.style.titleFont.weight,
-                fontFamily = FontFamily(
-                    Font(tile.style.titleFont.family)
-                ),
-                fontSize = tile.style.titleFont.size,
-                lineHeight = tile.style.titleFont.lineHeight,
-                textAlign = tile.style.titleFont.alignment,
-            )
-        }
-    }
-}
-
-@Composable
-fun Footer(footer: Footer) {
-    val foregroundColor = if (isLight(footer.style.backgroundColor)) Color.Black else Color.White // ktlint-disable max-line-length
-    val buttonForegroundColor = if (isLight(footer.style.buttonBackgroundColor)) Color.Black else Color.White // ktlint-disable max-line-length
-    Card(
-        shape = RoundedCornerShape(footer.layout.cornerRadius)
-    ) {
-        Column(
-            verticalArrangement = Arrangement.spacedBy(
-                footer.layout.verticalSpacing
-            ),
-            modifier = Modifier
-                .background(footer.style.backgroundColor)
-                .padding(
-                    top = footer.layout.padding,
-                    start = footer.layout.padding,
-                    end = footer.layout.padding,
-                    bottom = footer.layout.padding
-                ),
-        ) {
-            Text(
-                text = footer.model.title,
-                color = foregroundColor,
-                overflow = TextOverflow.Ellipsis,
-                fontWeight = footer.style.titleFont.weight,
-                fontFamily = FontFamily(
-                    Font(footer.style.titleFont.family)
-                ),
-                fontSize = footer.style.titleFont.size,
-                lineHeight = footer.style.titleFont.lineHeight,
-                textAlign = footer.style.titleFont.alignment,
-            )
-            Text(
-                text = footer.model.description,
-                color = foregroundColor,
-                overflow = TextOverflow.Ellipsis,
-                fontWeight = footer.style.descriptionFont.weight,
-                fontFamily = FontFamily(
-                    Font(footer.style.descriptionFont.family)
-                ),
-                fontSize = footer.style.descriptionFont.size,
-                lineHeight = footer.style.descriptionFont.lineHeight,
-                textAlign = footer.style.descriptionFont.alignment,
-            )
-            Column(
-                verticalArrangement = Arrangement.spacedBy(
-                    footer.layout.verticalSpacing
-                ),
-                horizontalAlignment = Alignment.CenterHorizontally,
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                Image(
-                    painter = painterResource(id = footer.model.imageId),
-                    contentDescription = null,
-                    contentScale = ContentScale.Fit,
-                    alignment = Alignment.Center
-                )
-                Button(
-                    onClick = { /*TODO*/ },
-                    modifier = Modifier.fillMaxWidth(),
-                    colors = ButtonDefaults.buttonColors(
-                        backgroundColor = footer.style.buttonBackgroundColor
-                    )
-                ) {
-                    Text(
-                        text = footer.model.buttonTitle,
-                        color = buttonForegroundColor,
-                        fontWeight = footer.style.buttonTitleFont.weight,
-                        fontFamily = FontFamily(
-                            Font(footer.style.buttonTitleFont.family)
-                        ),
-                        fontSize = footer.style.buttonTitleFont.size,
-                        lineHeight = footer.style.buttonTitleFont.lineHeight,
-                        textAlign = footer.style.buttonTitleFont.alignment,
-                    )
-                }
-            }
-        }
-    }
-}
-
-// DATA
-
-fun yourDataContainers(
-    tiles: List<TileModel>,
-    layout: ContainerLayout,
-    bigTileLayout: TileLayout,
-    bigTileStyle: TileStyle,
-    smallTileLayout: TileLayout,
-    smallTileStyle: TileStyle
-): List<Container> {
-    val tilesPerContainer = 3
-    val chunked = tiles.chunked(tilesPerContainer)
-
-    val containersConfig: List<ContainerType> = listOf(
-        ContainerType.LARGELEFT,
-        ContainerType.ROW,
-        ContainerType.LARGERIGHT,
-        ContainerType.ROW
-    )
-
-    return chunked.mapIndexed { index, tiles ->
-        when (val type = containersConfig[index % containersConfig.count()]) {
-            ContainerType.LARGELEFT -> Container(
-                type = type,
-                tiles = tiles.mapIndexed { index, tileModel ->
-                    if (index == 0) {
-                        Tile(
-                            tileModel,
-                            bigTileStyle,
-                            bigTileLayout,
-                            TileType.BIG
-                        )
-                    } else {
-                        Tile(
-                            tileModel,
-                            smallTileStyle,
-                            smallTileLayout,
-                            TileType.SMALL
-                        )
-                    }
-                },
-                layout = layout
-            )
-            ContainerType.ROW -> Container(
-                type = type,
-                tiles = tiles.map { tileModel ->
-                    Tile(
-                        tileModel,
-                        smallTileStyle,
-                        smallTileLayout,
-                        TileType.SMALL
-                    )
-                },
-                layout = layout
-            )
-            ContainerType.LARGERIGHT -> Container(
-                type = type,
-                tiles = tiles.mapIndexed { index, tileModel ->
-                    if (index == tiles.count() - 1) {
-                        Tile(
-                            tileModel,
-                            bigTileStyle,
-                            bigTileLayout,
-                            TileType.BIG
-                        )
-                    } else {
-                        Tile(
-                            tileModel,
-                            smallTileStyle,
-                            smallTileLayout,
-                            TileType.SMALL
-                        )
-                    }
-                },
-                layout = layout
-            )
-        }
-    }
-}
-
-fun rowContainers(
-    tiles: List<TileModel>,
-    tilesPerContainer: Int,
-    layout: ContainerLayout,
-    tileLayout: TileLayout,
-    tileStyle: TileStyle,
-    tileType: TileType
-): List<Container> {
-    val chunked = tiles.chunked(tilesPerContainer)
-
-    return chunked.map {
-        Container(
-            type = ContainerType.ROW,
-            layout = layout,
-            tiles = it.map { tileModel ->
-                Tile(
-                    tileModel,
-                    tileStyle,
-                    tileLayout,
-                    tileType
-                )
-            },
-        )
-    }
-}
-
 fun section(
     model: SectionModel,
     style: SectionStyle,
@@ -720,13 +248,13 @@ fun section(
         SectionType.DATA_KNOW_HOW -> return Section(
             model = model,
             type = model.type,
-            containers = rowContainers(
+            containers = yourDataContainers(
                 model.tiles,
-                tilesPerContainer = 3,
                 containerLayout,
+                bigTileLayout,
+                bigTileStyle,
                 smallTileLayout,
-                smallTileStyle,
-                TileType.SMALL
+                smallTileStyle
             ),
             layout = layout,
             style = style
